Ready4Hire/app/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LLM & AI
langchain>=0.0.350
langchain-community>=0.0.10
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2

# ML/DL
scikit-learn>=1.3.2
numpy>=1.24.0
umap-learn>=0.5.5
hdbscan>=0.8.33

# Audio
openai-whisper>=20231117
pyttsx3>=2.90

# Utilidades
langid>=1.1.6
python-dotenv>=1.0.0

# Base de datos (Async)
asyncpg>=0.29.0
aioredis>=2.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0  # Para tests de FastAPI

# Linting & Quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0
bandit>=1.7.5
safety>=2.3.5

# CI/CD & Monitoring
locust>=2.17.0  # Load testing
